fix: use safer default for docker port mapping
If the docker port mapping does not use a specific IP, it the port
will be opened for all available network devices. This might be
potentially dangerous, because people who just quickly want to try
out the DVWA are not always aware of this issue. Nevertheless anyone
on their local network could then access the DVWA inside the container.

Therefore this commit changes the default to be mapped to the same
port only on the local loopback device as the safer default, and adds
a short explanation in the respectice section in README, so that
people can make it available on all network devices, if they really
need to.